// src/model/viewModel/VwLoadDeliveryPlanBookingUser.ts
export interface VwLoadDeliveryPlanBookingUser {
  signTitle: string;
  codeTitile: string;
  fromTime: string;
  dateTime: string;
  isActive: boolean;
  capacity: number;
  maxCapacity: number;
  toTime: string;
  planBookingIsActive: boolean;
  remainCapacity: number;
  dname: string;
  planId: string;
  pdate: string;
}